requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
python-dotenv
langserve
dashscope